Automatically import packages
I made it through about three packages before I realized that importing
the transitive depenencies would be tedious. So thirty minutes and
ninety lines of Python later, we now have an automated importer. This is
a fairly simple implementation so care should be taken to look at the
generated diffs for now.